Use spl-token by default
