+++ associated-token-account/program/Cargo.toml
@@ -17,11 +17,11 @@ borsh = "0.9.1"
 num-derive = "0.3"
 num-traits = "0.2"
 solana-program = "1.9.5"
+spl-token = { version = "3.3", path = "../../token/program", features = ["no-entrypoint"] }
 spl-token-2022 = { version = "0.2", path = "../../token/program-2022", features = ["no-entrypoint"] }
 thiserror = "1.0"
 
 [dev-dependencies]
-spl-token = { version = "3.3", path = "../../token/program", features = ["no-entrypoint"] }
 solana-program-test = "1.9.5"
 solana-sdk = "1.9.5"
