+++ associated-token-account/program/tests/extended_mint.rs
@@ -13,7 +13,7 @@ use {
         transaction::{Transaction, TransactionError},
     },
     spl_associated_token_account::{
-        get_associated_token_address, instruction::create_associated_token_account,
+        get_associated_token_address_with_program_id, instruction::create_associated_token_account,
     },
     spl_token_2022::{
         error::TokenError,
@@ -96,10 +96,16 @@ async fn test_associated_token_account_with_transfer_fees() {
     transaction.sign(&[&payer], recent_blockhash);
     banks_client.process_transaction(transaction).await.unwrap();
 
-    let associated_token_address_sender =
-        get_associated_token_address(&wallet_address_sender, &token_mint_address);
-    let associated_token_address_receiver =
-        get_associated_token_address(&wallet_address_receiver, &token_mint_address);
+    let associated_token_address_sender = get_associated_token_address_with_program_id(
+        &wallet_address_sender,
+        &token_mint_address,
+        &spl_token_2022::id(),
+    );
+    let associated_token_address_receiver = get_associated_token_address_with_program_id(
+        &wallet_address_receiver,
+        &token_mint_address,
+        &spl_token_2022::id(),
+    );
 
     // mint tokens
     let sender_amount = 50 * maximum_fee;
