+++ associated-token-account/program/tests/process_create_associated_token_account.rs
@@ -12,20 +12,20 @@ use {
         transaction::{Transaction, TransactionError},
     },
     spl_associated_token_account::{
-        get_associated_token_address, instruction::create_associated_token_account,
+        get_associated_token_address_with_program_id, instruction::create_associated_token_account,
     },
     spl_token_2022::{extension::ExtensionType, state::Account},
 };
-
-#[allow(deprecated)]
-use spl_associated_token_account::create_associated_token_account as deprecated_create_associated_token_account;
 
 #[tokio::test]
 async fn test_associated_token_address() {
     let wallet_address = Pubkey::new_unique();
     let token_mint_address = Pubkey::new_unique();
-    let associated_token_address =
-        get_associated_token_address(&wallet_address, &token_mint_address);
+    let associated_token_address = get_associated_token_address_with_program_id(
+        &wallet_address,
+        &token_mint_address,
+        &spl_token_2022::id(),
+    );
 
     let (mut banks_client, payer, recent_blockhash) =
         program_test(token_mint_address, true).start().await;
@@ -71,8 +71,11 @@ async fn test_associated_token_address() {
 async fn test_create_with_fewer_lamports() {
     let wallet_address = Pubkey::new_unique();
     let token_mint_address = Pubkey::new_unique();
-    let associated_token_address =
-        get_associated_token_address(&wallet_address, &token_mint_address);
+    let associated_token_address = get_associated_token_address_with_program_id(
+        &wallet_address,
+        &token_mint_address,
+        &spl_token_2022::id(),
+    );
 
     let (mut banks_client, payer, recent_blockhash) =
         program_test(token_mint_address, true).start().await;
@@ -128,8 +131,11 @@ async fn test_create_with_fewer_lamports() {
 async fn test_create_with_excess_lamports() {
     let wallet_address = Pubkey::new_unique();
     let token_mint_address = Pubkey::new_unique();
-    let associated_token_address =
-        get_associated_token_address(&wallet_address, &token_mint_address);
+    let associated_token_address = get_associated_token_address_with_program_id(
+        &wallet_address,
+        &token_mint_address,
+        &spl_token_2022::id(),
+    );
 
     let (mut banks_client, payer, recent_blockhash) =
         program_test(token_mint_address, true).start().await;
@@ -185,8 +191,11 @@ async fn test_create_with_excess_lamports() {
 async fn test_create_account_mismatch() {
     let wallet_address = Pubkey::new_unique();
     let token_mint_address = Pubkey::new_unique();
-    let _associated_token_address =
-        get_associated_token_address(&wallet_address, &token_mint_address);
+    let _associated_token_address = get_associated_token_address_with_program_id(
+        &wallet_address,
+        &token_mint_address,
+        &spl_token_2022::id(),
+    );
 
     let (mut banks_client, payer, recent_blockhash) =
         program_test(token_mint_address, true).start().await;
@@ -253,8 +262,11 @@ async fn test_create_account_mismatch() {
 async fn test_create_associated_token_account_using_legacy_implicit_instruction() {
     let wallet_address = Pubkey::new_unique();
     let token_mint_address = Pubkey::new_unique();
-    let associated_token_address =
-        get_associated_token_address(&wallet_address, &token_mint_address);
+    let associated_token_address = get_associated_token_address_with_program_id(
+        &wallet_address,
+        &token_mint_address,
+        &spl_token_2022::id(),
+    );
 
     let (mut banks_client, payer, recent_blockhash) =
         program_test(token_mint_address, true).start().await;
@@ -284,53 +296,6 @@ async fn test_create_associated_token_account_using_legacy_implicit_instruction(
     create_associated_token_account_ix
         .accounts
         .push(AccountMeta::new_readonly(sysvar::rent::id(), false));
-
-    let mut transaction =
-        Transaction::new_with_payer(&[create_associated_token_account_ix], Some(&payer.pubkey()));
-    transaction.sign(&[&payer], recent_blockhash);
-    banks_client.process_transaction(transaction).await.unwrap();
-
-    // Associated account now exists
-    let associated_account = banks_client
-        .get_account(associated_token_address)
-        .await
-        .expect("get_account")
-        .expect("associated_account not none");
-    assert_eq!(associated_account.data.len(), expected_token_account_len);
-    assert_eq!(associated_account.owner, spl_token_2022::id());
-    assert_eq!(associated_account.lamports, expected_token_account_balance);
-}
-
-#[tokio::test]
-async fn test_create_associated_token_account_using_deprecated_instruction_creator() {
-    let wallet_address = Pubkey::new_unique();
-    let token_mint_address = Pubkey::new_unique();
-    let associated_token_address =
-        get_associated_token_address(&wallet_address, &token_mint_address);
-
-    let (mut banks_client, payer, recent_blockhash) =
-        program_test(token_mint_address, true).start().await;
-    let rent = banks_client.get_rent().await.unwrap();
-    let expected_token_account_len =
-        ExtensionType::get_account_len::<Account>(&[ExtensionType::ImmutableOwner]);
-    let expected_token_account_balance = rent.minimum_balance(expected_token_account_len);
-
-    // Associated account does not exist
-    assert_eq!(
-        banks_client
-            .get_account(associated_token_address)
-            .await
-            .expect("get_account"),
-        None,
-    );
-
-    // Use legacy instruction creator
-    #[allow(deprecated)]
-    let create_associated_token_account_ix = deprecated_create_associated_token_account(
-        &payer.pubkey(),
-        &wallet_address,
-        &token_mint_address,
-    );
 
     let mut transaction =
         Transaction::new_with_payer(&[create_associated_token_account_ix], Some(&payer.pubkey()));
